Show identity info in node details, re-verification
The new node ids are used to identify nodes. Settings dialog
is connected to Settings, and if a new node is added to pool
while the dialog is open, it will show up. The same goes for
re-verification (the node's certificate will be updated without
reopening the settings dialog).

Node edit dialog also has a button to trigger a re-verification,
which will be used when the peers identity changes and sync
doesn't work anymore. Instead of re-adding the peer, the user
can verify the new identity in this way. It opens the same wizard
as when adding a new node, except there's not the search page.