Info.plist,UX> [macOS] Fix missing focus on a dialog when LSUIElement=1
Using the LSUIElement key resulted in the dialogs (SettingsDialog, AboutDialog) doesn't obtain focus (doesn’t become the key window) when opened (through the HaveClip's menu in the system statusbar).

Just to be clear, dialogs are not overlapped by other windows when opened, but the focus remains on a previous window.

References:
- https://stackoverflow.com/questions/11160180/forcing-an-lsuielement-accessory-application-to-the-front
- https://bugreports.qt.io/browse/QTBUG-28327
- open() -> show() -> setVisible(true) https://code.qt.io/cgit/qt/qtbase.git/tree/src/plugins/platforms/cocoa/qcocoawindow.mm?h=v5.12.8#n363
- raise()                              https://code.qt.io/cgit/qt/qtbase.git/tree/src/plugins/platforms/cocoa/qcocoawindow.mm?h=v5.12.8#n902

Note: The fix relates to the direct opening of dialogs through the HaveClip's menu (in the system statusbar) and does not affect the indirect opening of "SecurityCodePrompt.ui" (here, the behavior remains the same — the focus is not stolen but the window is still topmost).